add src/logs
